app/requirements.txt
Flask==2.1.2
numpy==1.22.4
requests==2.28.0
Pillow==9.1.1